+++ packages/react-libremap-components/cypress.config.ts
@@ -4,6 +4,6 @@ import { defineConfig } from 'cypress';
 export default defineConfig({
 	component: nxComponentTestingPreset(__filename, {
 		bundler: 'vite',
-		buildTarget: 'react-maplibre:build',
+		buildTarget: 'react-libremap-components:build',
 	}),
 });